Refactor: Move the COM init wrapper to a separate file
