+++ src/windows/com_init_wrapper.hpp
@@ -1,0 +1,38 @@
+#pragma once
+
+#include <windows.h>
+
+// A wrapper around COM library initialization. Calls CoInitializeEx in the
+// constructor and CoUninitialize in the destructor.
+class com_init_wrapper {
+public:
+  com_init_wrapper(DWORD dwCoInit) {
+    // We can safely continue as long as COM was either successfully
+    // initialized or already initialized.
+    // RPC_E_CHANGED_MODE means that CoInitializeEx was already called with
+    // a different concurrency model.
+    switch (CoInitializeEx(nullptr, dwCoInit)) {
+    case S_OK:
+    case S_FALSE:
+      m_initialized = true;
+      break;
+    }
+  }
+
+  ~com_init_wrapper() {
+    if (m_initialized) {
+      CoUninitialize();
+      m_initialized = false;
+    }
+  }
+
+  com_init_wrapper(const com_init_wrapper &other) = delete;
+  com_init_wrapper &operator=(const com_init_wrapper &other) = delete;
+  com_init_wrapper(com_init_wrapper &&other) = delete;
+  com_init_wrapper &operator=(com_init_wrapper &&other) = delete;
+
+  bool is_initialized() const { return m_initialized; }
+
+private:
+  bool m_initialized = false;
+};
