+++ src/windows/webview_msedge.h
@@ -95,41 +95,6 @@ std::wstring get_file_version_string(const std::wstring &file_path) noexcept {
   return std::wstring(version, version_length);
 }
 
-// A wrapper around COM library initialization. Calls CoInitializeEx in the
-// constructor and CoUninitialize in the destructor.
-class com_init_wrapper {
-public:
-  com_init_wrapper(DWORD dwCoInit) {
-    // We can safely continue as long as COM was either successfully
-    // initialized or already initialized.
-    // RPC_E_CHANGED_MODE means that CoInitializeEx was already called with
-    // a different concurrency model.
-    switch (CoInitializeEx(nullptr, dwCoInit)) {
-    case S_OK:
-    case S_FALSE:
-      m_initialized = true;
-      break;
-    }
-  }
-
-  ~com_init_wrapper() {
-    if (m_initialized) {
-      CoUninitialize();
-      m_initialized = false;
-    }
-  }
-
-  com_init_wrapper(const com_init_wrapper &other) = delete;
-  com_init_wrapper &operator=(const com_init_wrapper &other) = delete;
-  com_init_wrapper(com_init_wrapper &&other) = delete;
-  com_init_wrapper &operator=(com_init_wrapper &&other) = delete;
-
-  bool is_initialized() const { return m_initialized; }
-
-private:
-  bool m_initialized = false;
-};
-
 #include "mswebview_engine.hpp"
 
 // Enable built-in WebView2Loader implementation by default.
